requirements.txt
Django>=4.2.0,<5.0
Pillow>=10.0.0
python-dotenv>=1.0.0
django-allauth>=0.54.0
whitenoise>=6.5.0
gunicorn>=21.2.0
django-unfold>=0.72.0